Dev: Fix helper_spec.feature from failing after the first run
Because we manually include `ApplicationHelper` into the `_view`
object, we can't have conflicting method names between the helpers.

This is not really something to worry about with user projects: I
believe most projects use `helper :all` and/or do not utilize
inheritance (include helper in a helper, with conflicting names in
`ApplicationHelper`).

Fixes #554