Work towards improving rspec-support script support.
rspec-rails has been a little different than the other repos for a
while. As the shared script setups are improving rspec-rails should
evolve to leverage these as much as possible.

Copy current support function and clone scripts. Update the run build
script to leverage the new scripts. Along the same lines, tweak the
travis build to be more inline as well.