Update Changelog with `rails g` changes in #1248.
TestUnit gets special treatment when it comes to the generators listed
when running `rails generate`. Since it is not set as the test framework
until much later in the startup process all of it's generators can be
seen. Unfortunately, when RSpec hooks into the Railtie we don't have the
option of deferring until later. This hid all of the available RSpec
generators.

This change allows users to see and run the default generators
standalone. While this isn't a common practice, it is sometimes useful
when certain components are manually created outside of
`rails g COMPONENT`.

[ci skip]