Add specs around existing settings
These specs are in a more integration style; instead of a more
unit-collaborator approach which would simply ensure that the proper
messages were sent to the configuration object.

The reason is that users of rspec-rails have specific expectations on
how the configuration options are available. If rspec-core changes
anything related to these calls that will break the rspec-rails user's
expectations. To ensure that doesn't happen we are fully testing the
integration between our shim and the `RSpec.configuration`.