Add explicit git master web console dependency
This is done because rails 5 needs a newer version of rails console than
will be installed by the implicit dependencies resolved via rubygems. It
is also only added to the development group because we do not want
webconsole to get in the way of any of the specs that we run.