Substitute quotes with underscores for method name
Rails' SystemTestCase uses `method_name` to generate a filename for a screenshot, which is output along with the failure message. The filename for examples with single or double quotation marks retain the marks, making it tedious to copy-and-past the generated filename then manually escape it to use in a shell command.